Split default-extension installer tests
