+++ tests/install-stage1-test-helpers.mjs
@@ -1,4 +1,4 @@
-import { mkdtempSync } from "node:fs";
+import { existsSync, mkdtempSync, readFileSync } from "node:fs";
 import { tmpdir } from "node:os";
 import { join } from "node:path";
 
@@ -18,4 +18,16 @@ export function captureConsole(method, callback) {
 		console[method] = original;
 	}
 	return lines.join("\n");
+}
+
+export function readPiLog(path) {
+	if (!existsSync(path)) return [];
+	return readFileSync(path, "utf8").trim().split(/\r?\n/).filter(Boolean);
+}
+
+export function readPiLogRecords(path) {
+	return readPiLog(path).map((line) => {
+		const [agentDir, cwd, ...commandParts] = line.split("|");
+		return { agentDir, cwd, command: commandParts.join("|") };
+	});
 }
